ocr-saas/backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Celery and RabbitMQ
celery[redis]>=5.3.6
aio-pika>=9.3.0

# MinIO (S3 Storage)
minio>=7.2.0
aiofiles>=23.2.1

# vLLM (Model Serving)
vllm>=0.3.0
openai>=1.12.0

# ML/AI Libraries
torch>=2.1.0
transformers>=4.37.0
onnxruntime-gpu>=1.16.0; platform_system != "Darwin"
onnxruntime>=1.16.0; platform_system == "Darwin"
Pillow>=10.2.0
pdf2image>=1.16.3
pypdfium2>=4.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3
structlog>=24.1.0
prometheus-client>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
ruff>=0.1.0
mypy>=1.8.0

# Image processing
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
scipy>=1.12.0
pytesseract>=0.3.10
